Fixes and updates to graph rendering
